optimize myst_mman_pids_munmap()

+++ include/myst/mmanutils.h
@@ -67,8 +67,8 @@ int proc_pid_maps_vcallback(myst_buf_t* vbuf);
 /* marks all pages in the mapping as owned by the given process */
 int myst_mman_pids_set(const void* addr, size_t length, pid_t pid);
 
-/* test whether all pages in the mapping are owned by the given process */
-int myst_mman_pids_test(const void* addr, size_t length, pid_t pid);
+/* return the length in bytes that are owned by the given process */
+ssize_t myst_mman_pids_test(const void* addr, size_t length, pid_t pid);
 
 /* release all pages in the mapping that are owned by the given process */
 int myst_mman_pids_munmap(const void* addr, size_t length, pid_t pid);
